requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
requests>=2.31
